Restructure and amend documentation (Sphinx).
Previously all documentation was a single large README.rst. Now that we
have started to implement and document, by means of docstrings, the API,
Sphinx was added for its beautiful documentation rendering.

The opportunity was taken to simplify the README.rst so that it
explained how the project could be installed and the documentation being
built. All other and more extensive documentation was moved to the
Sphinx based documentation in the ``docs`` folder.