Update dependencies to make regular install work.
Code generated by ``grpcio-tools`` depends on the ``protobuf`` package.
Rather obvious, as gRPC is an extention of Protobuf. However one would
have expected that ``grpcio`` package, being the gRPC runtime,  would
have a dependency on ``protobuf``. Alas, it does not not. Instead of
constantly figuring out which version of the ``protobuf`` package
matches which version of ``grpcio``, I have simply added a runtime
dependency on ``grpcio-tools`` that does have the proper ``protobuf``
dependency. That way I only have to worry about to correct
grpcio{,-tools} version number.

In addition I'v added ``colorama`` as as runtime dependency for colored
console output. The code assumed it was present, but ``setup.cfg``
didn't specify it for the regular install.