Add cmdline utilities for managing Orchestor ports.
- add
- list
- delete
- enable
- disable

Added option `with_scheduler` to `init_app` as all these cmdline
utitilies do not need a scheduler initiliazed and running.

Changed the `logger.exception` in `db_session` to `logger.info`.
Sometimes we rely on an exception to be thrown (EAFP) and then we
don't want to be confronted with a stack trace. As the exception is
reraised there is still the opportunity to have the stack traced logged
if we choose so.

EAFP: https://docs.python.org/3/glossary.html#term-eafp