Extend `Reservation` model with `{src,dst}_selected_vlan` column.
Reservations can be made using unqualified STPs. This means that a VLAN
range has been specified instead of a specific VLAN. In that case, it is
up to SuPA to select an appropriate VLAN out of the specified range.

The selected VLAN needs to be administered somewhere. The `Connection`
model, a model that only stores a specific VLAN, will only be
instantiated if a corresponding connection has been created in the
NRM/Orchestrator. Hence, that is too late in the process. As we're still
dealing with the reservation phase, storing it in the `Reservation`
model seems obvious.

Also adjusted the column type of `{src,dst}_vlan)` in the `Connection`
model from `Text` to `Integer`. After all, a single VLAN is represented
as a number.