Document running doctest from pytest.
Pytest is configured to collect tests from the top level `tests`
directory.  This means that, even when configured in `setup.cfg` with:

   addopts=--doctest-modules

it will not pick up any doctests. After all they all reside in the top
level `src` directory (and below). Currently there does **not** seem to
be a way to have pytest both run:

- regular tests below `tests`
- and doctests below `src`

Hence we have updated our documentation with instructions to run a
different pytest command for running the doctests.

A warning for a third party library needed to be disabled for pytest to
continue processing doctests.