Rename state names (including case) to those of Protobuf version.
The Protobuf messages, representing the three state machines, copied the
names of the different states verbatim from the SOAP ones. In our Python
code we had simplified some of these. After all why prefix the
reservation state machine ones with `RESERVE_` and provisioning and
lifecycle statem machine ones **not** with respectively `PROVISIONING_`
and `LIFECYCLE_`. That's inconsistent. Instead we choose to leave out the
`RESERVE_` prefix and picked CamelCase for the names instead of all
uppercase.

However, now that we need to map our 'Python' states to those of
Protobuf, we are hampered by not having a 1-to-1 mapping of the states.
So we made them identical to the Protobuf version.