Fix bug where `vlan=` was missing from labels.
We parse STPs into `Stp` objects. The individual attributes of such an
object end up in separate columns in the database. For the labels part
we strip away the `"?vlan=` part. When reconstituting an STP from its
individual parts (by means of on `Stp` object) we need to add back what
we stripped away.