Be explicit about `State` names versus values.
The statemachine library documentation is somewhat light on content.
Hence the difference between `State` names and values wasn't immediately
clear. First argument to `State` is the name, whereas the second
argument is the `value`. If the value is unspecified it would use, not
the name, but the identifier of the `State` as the value. The identifier
being the class attirbute the `State` instance in assigned to. To
prevent confusion we now supply both name and value to `State`
instances.